pyproject.toml
[project]
name = "langgraph-react-agent"
version = "0.1.0"
description = "A simple ReAct Agent built with LangGraph"
authors = [
    { name = "Brian Godwin Lim", email = "brian@brianlim.xyz" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.6",
    "python-dotenv>=1.0.1",
    "gradio>=5.24.0",
    "mlflow>=2.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]

[tool.setuptools.package-dir]
agent = "agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]
